==> 2_split_into_trips.Rmd <==
---
title: "Split into trips and filter out incorrectly inferred parts of trips"
author: "Gemma Clucas"
date: "22nd February 2021"
output: github_document
---

```{r setup, message = FALSE}
library(maptools)
library(sf)
library(patchwork)
library(ggspatial)
library(purrr)
library(rgdal)
library(raster)
library(plyr)
library(tidyverse)
library(sp)
library(ggplot2)
library(knitr)
library(spdplyr)
library(geosphere)
library(slider)
library(modelr)
library(fs)
options(scipen=999)
```

```{r global-options, include=FALSE}
knitr::opts_chunk$set(warning=FALSE, message=FALSE)
```
Having crawled the tracks in step 1 of the analysis, here I split the tracks into individual foraging trips, using time spent in close proximity to Saunders Island as the break point between trips.

I also remove sections of the trips where the average speed was less than 1.2 kph for 3 hours. These long sections of low speed movement have been inserted into the data by crawl. 

The tags we used were programmed to turn off when they were dry, and so we do not have fixes for when the birds were stationary at the colony. It also seems like we often miss fixes when they are transiting to/from the colony. Therefore, crawl has inserted long periods (over a day in some cases) of low speed movement between the end of one foraging bout and the start of another, when the bird was most likely at the colony incubating. I expect this over-abundance of points near the colony would bias our habitat models, and so I have removed the points in these sections of the trips.

This has the downside that we do not know exactly when a trip started or finished, but the resulting approximate start and finish times for the trips seem much more reasonable than without removing these sections of the trips; previously the birds were supposedly spending almost no time on land, which we know is not the case for chick-rearing parents.


### Load map and plot
```{r warning=FALSE}
Seamask<-readOGR("Seamask.shp")
#SSI <- crop(Seamask, c(450000, 750000, -600000, -100000)) # the original values I used here were cropping the end of the tracks when I filtered for the points off land, so I increased the extent of this base map to prevent that
# NB you can find the maximal extent of Seamask with extent(Seamask)
SSI <- crop(Seamask, c(450000, 1095192, -795043.9, -100000))

#Re-project to Lambert Azimuthal Equal Area
SSI_laea<-spTransform(SSI, CRS=CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m"))

# convert to dataframe for use with ggplot2
SSI_laea@data$id = rownames(SSI_laea@data)
SSI_laea.points = fortify(SSI_laea, region="id")
SSI_laea.df = plyr::join(SSI_laea.points, SSI_laea@data, by="id")

# filter out only the polygons for the islands
SSI_laea.df <- SSI_laea.df %>% filter(hole == TRUE)

# plot
SSI_laea.df %>% 
  ggplot(aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill="grey") +
  geom_path(color="grey") +
  coord_equal() +
  theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "aliceblue"))
```

### Bathymetry  
Plotting this is super slow so I am not going to continue developing this code here. I probably will need it for the final figures though.
```{r}
# SSI_bath <- raster("ssi_geotif/full_ssi18a.tif")
# #plot(SSI_bath)
# crs(SSI_bath)
# SSI_bath_WGS84 <- projectRaster(SSI_bath, crs=crs("+init=epsg:4326"))
# plot(SSI_bath_WGS84, xlim=c(-27,-25), ylim=c(-58.5,-57))


# # convert to a df for plotting in two steps,
# # First, to a SpatialPointsDataFrame then to a conventional dataframe
# SSI_bath_WGS84_df <- rasterToPoints(SSI_bath_WGS84, spatial = TRUE) %>% 
#   data.frame() %>% 
#   rename(Depth = layer)
# # tidy up
# rm(SSI_bath, SSI_bath_WGS84)

# ggplot() +
#   geom_raster(data = SSI_bath_WGS84_df , aes(x = x, y = y, fill = Depth)) +
#   coord_fixed(xlim = c(-27,-25), ylim=c(-58.5,-57))

```



### Plotting the penguin track onto the map
This function plots the raw crawled track onto a basic map.
```{r}
plot_track <- function(x) {
  # plot
  ggplot() + 
    geom_polygon(data = SSI_laea.df, aes(x = long, y = lat, group = group), fill="grey50") +
    geom_path(data = SSI_laea.df, aes(x = long, y = lat, group = group), color="grey50") +
    coord_equal() +
    geom_point(data = x, aes(x = LON, y = LAT)) +
    theme_bw() +
    theme(panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "aliceblue"),
          legend.title = element_blank()) 
}

```

### Remove points that are over land
This function also removes points within 1km of land. I tried buffers of 500m, 1km, and 2km around the island and found that 1km seemed to split trips fairly well without oversplitting them. However, there are definitely still some instances when multiple trips are not being split properly. These I split by hand (details below).
```{r}
# first create 500m buffer around island
SSI_laea_buffer <- buffer(SSI_laea, width=-1000)

remove_points_on_land <- function(track) {
  # make the track spatial points df
  coordinates(track) <- ~LON + LAT
  # tell it that it's projected in LAEA centred on Saunders
  proj4string(track) <- CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
  # add new column to track object identifying whether the track is off the island
  track$off_island <- !is.na(over(track, SSI_laea_buffer))
  # filter the points for just those that are off the island
  track %>%
    filter(off_island == TRUE)
}

```

### Split into trips 

First create a function for offsetting the values in a column by 1 row (or more if you change ```shiftLen```). This is needed to later calculate the lag time between each point.
```{r}
# function for offsetting values by 1 row
rowShift <- function(x, shiftLen = 1L) {
    rr <- (1L + shiftLen):(length(x) + shiftLen)
    rr[rr<1] <- NA
    return(x[rr])
}

```

Then we calculate the lag time between each point. If there is a lag longer than 10 minutes (0.16 hours), then this is when the bird was on land (technically inside the buffer zone) and so we can use this to split the track into separate foraging trips.  

Each time there is a lag greater than 10 minutes, we get a ```Start_trip = TRUE``` in the ```at_sea``` spatial object, otherwise ```Start_trip = FALSE```. 

In order to record the final trip, I need to fudge the final line of ```at_sea``` and give it ```Start_Trip == TRUE``` so that I can use ```Start_row_indexes``` to plot this final trip.

The row indexes of where these ```TRUE``` values occur are then used to add a column that records the trip number in the dataframe.
```{r}
split_into_trips <- function(at_sea) {
  # offset the values in Time_since by 1 row 
  at_sea2 <- at_sea %>% 
    arrange(Time_absolute, desc(locType == 'o')) %>%        # arrange times so that observed locations come before predicted
    dplyr::distinct(Time_absolute, .keep_all = TRUE) %>%    # remove duplicated times, keeping first row i.e. the observed one
    mutate(lag1 = rowShift(Time_since, -1), 
                               # calculate differences between Time_since and each lag 
                               diff1 = Time_since - lag1,
                               # put track$Start_trip == TRUE where the diff1 is greater than x mins
                               # Start_trip = diff1 >= 0.5) # 30 minutes on land
                               Start_trip = diff1 >= 0.16) # 10 minutes on land
  # Start_trip = diff1 >= 0.25) # 15 minutes on land
  # change the "NA" at the beginning of the first trip to "TRUE"
  at_sea2$Start_trip[1] <- TRUE
  # change the final value of Start_trip to TRUE to signal this is actually the end of the last trip
  at_sea2[length(at_sea2), "Start_trip"] <- TRUE
  # record where each trip begins
  Start_row_indexes <- as.list(which(at_sea2$Start_trip == TRUE))
  # make a sequence from 1 to the number of trips
  seq <- c(1:(length(Start_row_indexes) -1))
  # add a column that records the trip number in the dataframe
  at_sea2$Trip <-
    # this code takes the numbers in seq and repeats each for the number of rows between 
    # the start of trip n+1 and trip n (i.e. the length of the current trip)
    purrr::map(seq, ~rep(.x, each = (Start_row_indexes[[.x + 1]] - (Start_row_indexes[[.x]])))) %>% 
    unlist() %>%    # the result has to be changed from a list to a vector
    append(., values = tail(., n=1))    # and this repeats the last trip number once more and appends to the vector, otherwise it missed the last row
  return(at_sea2)
}


```



### Estimating and plotting average speed
I am estimating the average speed over a sliding window of 3 hours, so that I'm not filtering out times when the birds are resting between dives. 

I am grouping the dataframe by Trip, so that the average speed calculation is limited to points within a single trip. This is a little tricky as the sliding window from ```slide_dbl``` does not play well with ```group_by```, which is why there is a separate function called ```single_group_avg_speed``` that is then called inside the main function ```calc_avg_speed```. I found the code to do that [here](https://cran.r-project.org/web/packages/slider/vignettes/rowwise.html).

This takes the SpatialPointsDataFrame ```at_sea``` which has already been filtered to keep only points off the island.
```{r}
single_group_avg_speed <- function(.data_group, ...) {
  mean_speed <- slide_dbl(                      # slide_dbl() gives a dataframe as output
      .x = .data_group, 
      .f = ~mean(.$Speed_kph, na.rm = TRUE),    # the function to apply
      .before = 35,                             # this row and 35 before makes a 3 hour window
      .partial = TRUE                           # allows calculations on partial windows -> first 3 hours of the trip are not missed
    )
  mutate(.data_group, Avg_speed = mean_speed)
}


calc_avg_speed <- function(x){
  x2 <- x %>% 
    sp::spTransform(crs("+init=epsg:4326")) %>%
    data.frame() %>%
    dplyr::group_by(Trip) %>% 
    dplyr::mutate(Distance = distHaversine(p1 = cbind(LON, LAT),    # calculate distance between current and previous point in meters
                                           p2 = cbind(lag(LON), lag(LAT)),   # lag is a base r function that takes the next observation 
                                           r = 6362895)) %>%        # r = radius of earth in meters at 57.7 South
    dplyr::mutate(Speed_ms = Distance / (diff1*60*60))   %>%        # gives speed in m/s since diff1 is in decimal hours
    dplyr::mutate(Speed_kph = (Distance/1000) / diff1)  %>%         # speed in km per hour
    group_modify(single_group_avg_speed) %>%                        # calc average speed over 3 hour window within each trip
    ungroup() 
  # change it back to a spatialpointsdataframe with LAEA projection
  coordinates(x2) <- ~LON + LAT
  proj4string(x2) <- CRS("+proj=longlat +ellps=WGS84")
  x2 <- spTransform(x2, CRS = CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m"))
  # change to sf so that it's compatible with the tidyverse
  # x2 <- x2 %>% 
  #   st_as_sf()
  return(x2)
}

```


Function to plot average speed over time (ugly but fine for just eyeballing the data).
```{r}
plot_avg_speed <- function(x){
  x %>% 
    data.frame() %>%
    # Take every 36th row of the dataframe to speed up plotting
    dplyr::slice(which(row_number() %% 36 == 1)) %>%
    # Plot
    ggplot(aes(x=Time_absolute, y=Avg_speed, group = Ptt)) +                        
    geom_line() +
    xlab("") +
    scale_y_continuous(breaks=seq(0,15,0.5)) +
    # add a line at 1.2 kph to show where the cut-off will be
    geom_hline(yintercept=1.2,
               linetype="dashed",
               color = "red")
}

```


### Plot trips before trimming
I want to be able to view the trips before and after trimming sections and removing duff trips, to make sure nothing goes wrong.

This function takes a sequence of numbers corresponding to the number of ```Start_trip == TRUE``` in the object ```at_sea``` to split the dataframe up into trips and plot each indiviudally. The function will resize the map according to the min and max longitude and latitude for each trip, while still including Saunders Island.

Note, now that I am recording the trip number in ```at_sea``` I could use facting to plot the trips instead, probably. EDIT: Nope tried this, the graphs are too squashed.
```{r}
plot_trip_raw <- function(x) {
  ggplot() +
    # plot the map first
    geom_polygon(data = SSI_laea.df, aes(x = long, y = lat, group = group), fill="grey50") +
    geom_path(data = SSI_laea.df, aes(x = long, y = lat, group = group), color="grey50") +
    coord_equal() +
    # add the points. This uses the trip number (x) to subset the dataframe by trip.
    geom_point(data = data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1), ]), aes(x = LON, y = LAT, colour = Avg_speed)) +
    theme_bw() +
    # add headings for trip number and start/stop time
    labs(color="Average speed",
         title = paste0("Trip ", x),
         subtitle = paste0("Trip start: ",
                            data.frame(at_sea[Start_row_indexes[[x]], ]) %>% select(Time_absolute),
                            ", Trip end: ",
                            data.frame(at_sea[Start_row_indexes[[x+1]]-1, ]) %>% select(Time_absolute),
                           "\n # Observed locations: ", data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1), ]) %>%
                             filter(locType == "o") %>% nrow())) +
    theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "aliceblue")) +
    # size according to the dimensions of the trip
    # Saunders is always plotted by taking the min/max value of the island or the trip points for specifying the plot dimensions
    coord_fixed(ratio = 1,
                xlim = c(min(c(-33000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LON) %>% min())),
                         max(c(-22000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LON) %>% max()))),
                ylim = c(min(c(18000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LAT) %>% min())),
                         max(c(28000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LAT) %>% max()))),
                expand = TRUE,
                clip = "on")
}
```
  


### Filter out low speed sections of the trips
This removes low speed sections of trips and adds new labels in ```Start_trip_new``` to signal the new start of the trip. 

```{r}
filter_out_low_speed_sections <- function(x){
  x2 <- x %>% 
    data.frame() %>% 
    # remove points where the avg speed over the previous 3 hours was less than 1.2 kph
    dplyr::filter(Avg_speed > 1.2) %>%      
    group_by(Trip) %>%
    # create new flags for the start of the trip
    mutate(Start_trip_new = ifelse(Time_absolute == min(Time_absolute), TRUE, FALSE)) %>%    
    ungroup() 
  # fudge the final line so that it also says Start_trip_new = TRUE
  x2[nrow(x2), "Start_trip_new"] <- TRUE
  # change it back to a spatialpointsdataframe with LAEA projection
  coordinates(x2) <- ~LON + LAT
  proj4string(x2) <- CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
  return(x2)
}
```


### Plotting the trips after removing low speed sections and duff trips
This is very similar to above but uses ```Start_trip_new``` as the flag for the beginning of each trip.

```{r}
plot_trips_trimmed <- function(x) {
  ggplot() +
    # plot the map first
    geom_polygon(data = SSI_laea.df, aes(x = long, y = lat, group = group), fill="grey50") +
    geom_path(data = SSI_laea.df, aes(x = long, y = lat, group = group), color="grey50") +
    coord_equal() +
    # add the points. This uses the trip number (x) to subset the dataframe by trip.
    geom_point(data = data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1), ]), aes(x = LON, y = LAT, colour = Avg_speed)) +
    theme_bw() +
    # add headings for trip number and start/stop time
    labs(color="Average speed",
         title = paste0("Trip ", x),
         subtitle = paste0("Trip start: ",
                           data.frame(at_sea[Start_row_indexes[[x]], ]) %>% select(Time_absolute),
                           ", Trip end: ",
                           data.frame(at_sea[Start_row_indexes[[x+1]]-1, ]) %>% select(Time_absolute),
                           "\n # Observed locations: ", data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1), ]) %>%
                             filter(locType == "o") %>% nrow())) +
    theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_rect(fill = "aliceblue")) +
    # size according to the dimensions of the trip
    # Saunders is always plotted by taking the min/max value of the island or the trip points for specifying the plot dimensions
    coord_fixed(ratio = 1,
                xlim = c(min(c(-33000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LON) %>% min())),
                         max(c(-22000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LON) %>% max()))),
                ylim = c(min(c(18000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LAT) %>% min())),
                         max(c(28000, data.frame(at_sea[c(Start_row_indexes[[x]]:Start_row_indexes[[x+1]]-1),]) %>% select(LAT) %>% max()))),
                expand = TRUE,
                clip = "on")
  
}

```



### Run through these steps for each penguin

The numbers of the Ptt devices go from 196697 to 196716.

Load crawled track.
```{r}
penguin <- "196716"
predObj <- read.csv(paste0("predicted_tracks/", penguin, "_track.csv"), stringsAsFactors = FALSE)

# select the useful columns and rename
track <- predObj %>%
  dplyr::select(Ptt, locType, Time_absolute, Time_since, mu.x, mu.y) %>%
  dplyr::rename(LON = mu.x, LAT = mu.y)

plot_track(track)
```


Remove points on land and (optionally) replot to make sure nothing has gone wrong.
```{r}
at_sea <- remove_points_on_land(track)
# at_sea %>% data.frame() %>% plot_track(.)
```


Split into trips.
```{r}
at_sea <- split_into_trips(at_sea)
# head(at_sea)  # check that the table starts with Start_trip = TRUE and trip number 1
# tail(at_sea)  # check that the table ends with Start_trip = TRUE and some higher trip number
```


Calculate the average speed in 3 hour windows
```{r}
at_sea <- calc_avg_speed(at_sea)

plot_avg_speed(at_sea)
```

Check that 1.2 kph looks like a good cut-off. 

Note that for Ptt 196700, I used 2.2 kph as there were long sections (days) of movement inferred at speeds closer to 2 kph, which cannot be real. Also for Ptt 196701 I used 1.6 kph as a cut-off. For Ptt 196706 I also used 2.2 kph.


Plot trips before doing any trimming.
```{r}
Start_row_indexes <- as.list(which(at_sea$Start_trip == TRUE))    # need this for plot_trips() function

plots <- at_sea %>%
  data.frame() %>%                  # convert to dataframe
  distinct(Trip) %>%                # find number of trips
  deframe() %>%                     # change to vector to pass to map
  purrr::map(., ~plot_trip_raw(.x))

invisible(lapply(plots, print))
```



Filter out the low speed sections of the trips, remove trips with 1 or 0 observed fixes, then plot. Check that all trips with fewer than 2 observed fixes have been removed and that none of the real trips are missing.
```{r, message=FALSE}
at_sea <- filter_out_low_speed_sections(at_sea) 

# Remove duff trips
Keep_trips <- at_sea %>% 
  data.frame() %>% 
  group_by(Trip) %>% 
  filter(locType == "o") %>% 
  count() %>% 
  filter(n > 1) %>% 
  pull(Trip) 
at_sea <- at_sea %>% filter(Trip %in% Keep_trips)

# add Start_trip_new == TRUE to the end of the dataframe in case the last trip was trimmed
at_sea[nrow(at_sea), "Start_trip_new"] <- TRUE

# need this for plot_trips() function
Start_row_indexes <- as.list(which(at_sea$Start_trip_new == TRUE))    

# create plots for each trip
plots <- at_sea %>%
  data.frame() %>%                  
  distinct(Trip) %>%                
  nrow() %>%                        # find number of trips
  seq(1, .) %>% 
  purrr::map(., ~plot_trips_trimmed(.x))

invisible(lapply(plots, print))

```

### Do any trips need to be split further?

If trips are not being split up properly because there were no fixes within the buffer zone around the island, use this code to plot the problematic trip(s) and  work out where the split should go. Change the number in ```trip_number``` for the trip that needs to be split, and add rows in increments, specified by ```a```, to visualise when the penguin returns to the colony and leaves again. 

```{r}
trip_number <- 12

plot_trip_to_split <- function(trip_number, a) {
  ggplot() +
  # plot the map first
  geom_polygon(data = SSI_laea.df, aes(x = long, y = lat, group = group), fill="grey50") +
  geom_path(data = SSI_laea.df, aes(x = long, y = lat, group = group), color="grey50") +
  coord_equal() +
  # add the points. This uses the trip number (x) to subset the dataframe by trip.
  geom_point(data = data.frame(at_sea[c(Start_row_indexes[[trip_number]]:(Start_row_indexes[[trip_number]] + a)), ]), aes(x = LON, y = LAT, colour = Avg_speed)) +
  theme_bw() +
  # add headings for trip number and start/stop time
  labs(color="Average speed",
       subtitle = paste0("Trip start: ",
                           data.frame(at_sea[Start_row_indexes[[trip_number]], ]) %>% dplyr::select(Time_absolute),
                           ", Trip end: ",
                           data.frame(at_sea[(Start_row_indexes[[trip_number]] + a), ]) %>% dplyr::select(Time_absolute))) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "aliceblue")) +
  # size according to the dimensions of the trip
  # Saunders is always plotted by taking the min/max value of the island or the trip points for specifying the plot dimensions
  coord_fixed(ratio = 1,
              xlim = c(min(c(-33000, data.frame(at_sea[c(Start_row_indexes[[trip_number]]:(Start_row_indexes[[trip_number]] + a)), ]) %>% select(LON) %>% min())),
                         max(c(-22000, data.frame(at_sea[c(Start_row_indexes[[trip_number]]:(Start_row_indexes[[trip_number]] + a)), ]) %>% select(LON) %>% max()))),
                ylim = c(min(c(18000, data.frame(at_sea[c(Start_row_indexes[[trip_number]]:(Start_row_indexes[[trip_number]] + a)), ]) %>% select(LAT) %>% min())),
                         max(c(28000, data.frame(at_sea[c(Start_row_indexes[[trip_number]]:(Start_row_indexes[[trip_number]] + a)), ]) %>% select(LAT) %>% max()))),
              expand = TRUE,
              clip = "on")
}

# Use these to figure out roughly where the trip should be split and then hone in on the correct line by altering a
plot_trip_to_split(trip_number, 50)
plot_trip_to_split(trip_number, 100)
plot_trip_to_split(trip_number, 200)
plot_trip_to_split(trip_number, 300)

```


Below I have recorded where I have split trips by hand for each penguin.
```{r}
# # 196697
# at_sea[(Start_row_indexes[[2]] + 269), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[2]] + 640), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[9]] + 291), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[11]] + 313), "Start_trip_new"] <- TRUE

# # 196698
# at_sea[(Start_row_indexes[[10]] + 1034), "Start_trip_new"] <- TRUE

# # 196699
# at_sea[(Start_row_indexes[[1]] + 188), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[2]] + 285), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[10]] + 210), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[10]] + 424), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 515), "Start_trip_new"] <- TRUE

# # 196700
# at_sea[(Start_row_indexes[[4]] + 231), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 155), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 450), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 711), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 992), "Start_trip_new"] <- TRUE

# # 196701
# at_sea[(Start_row_indexes[[3]] + 333), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 37), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[12]] + 289), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[12]] + 478), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[12]] + 687), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 297), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[15]] + 116), "Start_trip_new"] <- TRUE

# # 196702
# at_sea[(Start_row_indexes[[3]] + 22), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[3]] + 256), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 194), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 249), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[6]] + 269), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[6]] + 590), "Start_trip_new"] <- TRUE

# 196703
# Nothing needs to be split

# # 196704
# at_sea[(Start_row_indexes[[4]] + 63), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 124), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[7]] + 100), "Start_trip_new"] <- TRUE
# # The last trip is really messy so I'm going to leave it as is

# # 19705
# at_sea[(Start_row_indexes[[17]] + 255), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[17]] + 725), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[20]] + 230), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[20]] + 520), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[21]] + 170), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[25]] + 428), "Start_trip_new"] <- TRUE

# # 19706
# at_sea[(Start_row_indexes[[7]] + 160), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[8]] + 57), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[8]] + 224), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[9]] + 391), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[9]] + 498), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[9]] + 588), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[11]] + 125), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 35), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 125), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 170), "Start_trip_new"] <- TRUE

# # 196707
# at_sea[(Start_row_indexes[[3]] + 298), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 436), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 836), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[24]] + 114), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[25]] + 439), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[25]] + 532), "Start_trip_new"] <- TRUE

# # 196708
# at_sea[(Start_row_indexes[[4]] + 135), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[6]] + 98), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[10]] + 40), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[11]] + 37), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[11]] + 382), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 360), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 900), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[15]] + 87), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[16]] + 132), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 93), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 204), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 330), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[20]] + 149), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[21]] + 133), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[21]] + 289), "Start_trip_new"] <- TRUE

# # 196709
# at_sea[(Start_row_indexes[[10]] + 265), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 97), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 338), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 614), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[13]] + 885), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[16]] + 215), "Start_trip_new"] <- TRUE

# #196710
# at_sea[(Start_row_indexes[[3]] + 98), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[6]] + 336), "Start_trip_new"] <- TRUE

# # 196711
# at_sea[(Start_row_indexes[[12]] + 288), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[12]] + 616), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[14]] + 304), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[17]] + 101), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[17]] + 185), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 101), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 196), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 253), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[18]] + 409), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[25]] + 300), "Start_trip_new"] <- TRUE

# # 196712
# at_sea[(Start_row_indexes[[2]] + 287), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[2]] + 569), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 304), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 637), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[7]] + 407), "Start_trip_new"] <- TRUE

# # 196713
# at_sea[(Start_row_indexes[[1]] + 349), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[2]] + 379), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[3]] + 213), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[3]] + 563), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[8]] + 390), "Start_trip_new"] <- TRUE

# # 196714
# at_sea[(Start_row_indexes[[1]] + 82), "Start_trip_new"] <- TRUE

# # 196715
# at_sea[(Start_row_indexes[[3]] + 334), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 287), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[4]] + 585), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 329), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 555), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[5]] + 878), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[9]] + 391), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[10]] + 407), "Start_trip_new"] <- TRUE
# at_sea[(Start_row_indexes[[11]] + 77), "Start_trip_new"] <- TRUE

# 196716
at_sea[(Start_row_indexes[[6]] + 69), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[12]] + 253), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[14]] + 377), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[14]] + 588), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[18]] + 200), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[18]] + 361), "Start_trip_new"] <- TRUE
at_sea[(Start_row_indexes[[26]] + 92), "Start_trip_new"] <- TRUE
```

Next, correct the trip numbers in the dataframe and replot the trips to make sure it's worked (don't need to change anything for each penguin here, just run).
```{r}
# update where each trip begins
Start_row_indexes <- as.list(which(at_sea$Start_trip_new == TRUE))
# make a sequence from 1 to the new number of trips
seq <- c(1:(length(Start_row_indexes) -1))
# correct the trip number in the dataframe
at_sea$Trip <-
  # this code takes the numbers in seq and repeats each for the number of rows between 
  # the start of trip n+1 and trip n (i.e. the length of the current trip)
  purrr::map(seq, ~rep(.x, each = (Start_row_indexes[[.x + 1]] - (Start_row_indexes[[.x]])))) %>% 
  unlist() %>%    # the result has to be changed from a list to a vector
  append(., values = tail(., n=1))    # and this repeats the last trip number once more and appends to the vector, otherwise it missed the last row


# create plots for each trip
plots <- at_sea %>%
  data.frame() %>%                  
  distinct(Trip) %>%                
  nrow() %>%                        # find number of trips
  seq(1, .) %>% 
  purrr::map(., ~plot_trips_trimmed(.x))

invisible(lapply(plots, print))

```



If everything seems to have worked, save the dataframe for each penguin and write to CSV.
```{r}
# Just save as dataframe
# assign(paste0("df_", penguin), data.frame(at_sea))

# Save and write to csv
# assign(paste0("df_", penguin), data.frame(at_sea)) %>%
#   write.csv(., paste0("Trimmed_trips/", penguin, "_trips.csv", sep = ""), row.names = FALSE)

```

To plot a single trip:
```{r}
# trip <- 6
# 
# ggplot() +
#   # plot the map first
#   geom_polygon(data = SSI_laea.df, aes(x = long, y = lat, group = group), fill="grey50") +
#   geom_path(data = SSI_laea.df, aes(x = long, y = lat, group = group), color="grey50") +
#   coord_equal() +
#   # add the points. This uses the trip number (x) to subset the dataframe by trip.
#   geom_point(data = at_sea %>% data.frame() %>% filter(Trip == trip), aes(x = LON, y = LAT, colour = Avg_speed)) +
#   coord_fixed(ratio = 1,
#               xlim = c(min(c(-33000, at_sea %>% data.frame() %>% filter(Trip == trip) %>% select(LON) %>% min())),
#                        max(c(-22000, at_sea %>% data.frame() %>% filter(Trip == trip) %>% select(LON) %>% max()))),
#               ylim = c(min(c(18000, at_sea %>% data.frame() %>% filter(Trip == trip) %>% select(LAT) %>% min())),
#                        max(c(28000, at_sea %>% data.frame() %>% filter(Trip == trip) %>% select(LAT) %>% max()))),
#               expand = TRUE,
#               clip = "on")

```



Merge the data from all individuals and save the merged dataframe to a csv file. 

```{r}

# data_dir <- "Trimmed_trips"
# 
# data_dir %>% 
#   dir_ls() %>%              # list all files in the directory
#   map_dfr(read_csv) %>%     # use purrr::map_dfr to read all CSVs into a single dataframe
#   select(-optional) %>%     # remove unnecessary columns
#   select(-optional.1) %>%
#   select(-optional.2)  %>% 
#   select(-Start_trip) %>% 
#   write.csv(., file = "Trimmed_trips/All_trimmed_trips.csv", row.names = FALSE)

```
