==> 6_Predictions.Rmd <==
---
title: "Making predictions from the model"
author: "Gemma Clucas"
date: "5/3/2021"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(raster)
library(rgdal)
library(ncdf4)
library(viridis)
library(mgcv)
library(ggplot2)
library(tidyverse)
```


Read in the data.
```{r}
data <- read.csv(file = "PresBackgroundLocationsWithEnvironmentalVariables.csv", stringsAsFactors = FALSE)
```

Run the final model.
```{r}
GAM <- gam(pres ~ s(colonydist, k = 3) + s(SST, k = 6), data=data, bs="cs", family=binomial, select=TRUE, method='GCV.Cp')

```


### Load map for plotting

Note this is unprojected (WGS84).
```{r}
Seamask<-readOGR("Seamask.shp")
SSI_polygons <- crop(Seamask, c(450000, 1095192, -695043.9, -100000)) %>% 
  spTransform(., crs("+init=epsg:4326"))

# convert to dataframe for use with ggplot2
SSI_polygons@data$id = rownames(SSI_polygons@data)
SSI_polygons.points = fortify(SSI_polygons, region="id")
SSI_polygons.df = plyr::join(SSI_polygons.points, SSI_polygons@data, by="id")

# filter out only the polygons for the islands
SSI_polygons.df <- SSI_polygons.df %>% filter(hole == TRUE)

# plot
SSI_polygons.df %>% 
  ggplot(aes(x = long, y = lat, group = group)) + 
  geom_polygon(fill="grey") +
  geom_path(color="grey") +
  theme_bw() +
  coord_fixed() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "aliceblue"))
```

## MPA boundaries

These are all individual shape files that I have downloaded from [here](https://add.data.bas.ac.uk/repository/entry/show?entryid=7f3136e7-8b50-4909-bb82-f4223a4f9768). There is a useful [viewer](https://sggis.gov.gs/) too.
They are in some weird projection so I am converting all to WGS84 before plotting.

```{r}
# function to convert SpatialPolygonsDataFrame to regular dataframe
spatialpolygons_to_df <- function(sp) {
  sp@data$id = rownames(sp@data)
  sp.points = fortify(sp, region="id")
  sp.df = plyr::join(sp.points, sp@data, by="id")
}

# # This is the entire extent of the MPA around SG and SSI
# SGSSI_MPA <- readOGR("MPA/SG_MPA/sg_mpa.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(SGSSI_MPA)
# SGSSI_MPA.df <- spatialpolygons_to_df(SGSSI_MPA)

# # CCAMLR management areas
# CCAMLR <- readOGR("MPA/Ccamlr_zones/Ccamlr_zones.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(CCAMLR)
# CCAMLR.df <- spatialpolygons_to_df(CCAMLR)

# 50 km no-take zone around SSI
NoTake_50km <- readOGR("MPA/sg_mpa_notake_ssi50km/sg_mpa_notake_ssi50km.shp") %>% 
  spTransform(., crs("+init=epsg:4326"))
#plot(NoTake_50km)
NoTake_50km.df <- spatialpolygons_to_df(NoTake_50km)

# # 50 km no-take zone around SSI trench
# NoTake_Trench50km <- readOGR("MPA/sg_mpa_notake_ssitrench50km/sg_mpa_notake_ssitrench50km.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(NoTake_Trench50km)
# NoTake_Trench50km.df <- spatialpolygons_to_df(NoTake_Trench50km)

# # No-take zone south of 60 South
# NoTake_60South <- readOGR("MPA/sg_mpa_notake_s60s/sg_mpa_notake_s60s.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(NoTake_60South)
# NoTake_60South.df <- spatialpolygons_to_df(NoTake_60South)

# # Pelagic closed areas (just SSI)
# Pelagic_closed <- readOGR("MPA/sg_mpa_pelagic_closed_areas/sg_mpa_pelagic_closed_areas.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(Pelagic_closed)
# Pelagic_closed.df <- spatialpolygons_to_df(Pelagic_closed)

# # Benthic closed areas (includes SG, not just SSI)
# Benthic_closed <- readOGR("MPA/sg_mpa_benthic_closed_areas/sg_mpa_benthic_closed_areas.shp") %>% 
#   spTransform(., crs("+init=epsg:4326"))
# #plot(Benthic_closed)
# Benthic_closed.df <- spatialpolygons_to_df(Benthic_closed)



```


## 1. Predict around Saunders to check that the distribution looks ok

This just creates a raster across the study area using the bathymetry raster cropped to the right extent.
```{r}
# Read in bathymetry raster and crop to extent of the study area around Saunders
SSI_bath_WGS84 <- raster("ssi_geotif/full_ssi18a.tif") %>% 
  projectRaster(., crs=crs("+init=epsg:4326")) %>% 
  crop(., c(-27.96598, -24.623, -58.41806, -57))

x <- SSI_bath_WGS84

# Project to LAEA to find extent in metres
y <- projectRaster(x, crs = CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m"))
extent(y)
```


Crop out land from the bathymetry raster using ```mask()``` 
```{r}
# Read in shapefile for land
SSI_WGS84 <- readOGR("Seamask.shp") %>% 
  crop(., c(450000, 1095192, -695043.9, -100000)) %>% 
  spTransform(., crs("+init=epsg:4326"))

# Cut out land
mask <- mask(x, SSI_WGS84, inverse=F)


# Plot to check
# plot(mask, col=viridis(100))
# plot(mask, xlim = c(-26.6, -26.2), ylim = c(-57.9, -57.7), col=viridis(100))
```




### Make a raster for the distance from colony for Saunders

```{r}
colony_lat<- -57.808 
colony_lon<- -26.404

# Find the colony cell in the study area raster
j <- cellFromXY(mask, cbind(colony_lon, colony_lat))
# Change the value of the cell where the colony is to 2 (all the other cells are 1)
mask[j]<-2 


# Create a distance raster from the colony
# Moving through land is prevented by omiting cells with NA values
dist <- gridDistance(mask, origin=2, omit=NA)
plot(dist, col=viridis(100))
#plot(dist, xlim = c(-26.6, -26.2), ylim = c(-57.9, -57.7), col=viridis(100))
```

### SST raster

Get the sea surface temperature data that I downloaded previously.
```{r}
# Open data - note there is a dataset for Jan and one for Feb
nc_data_Jan <- nc_open('global-analysis-forecast-phy-001-024-monthly_January2020.nc')
nc_data_Feb <- nc_open('global-analysis-forecast-phy-001-024-monthly_February2020.nc')
# Extract longs and lats
nc_lon <- ncvar_get(nc_data_Jan, "longitude")
nc_lat <- ncvar_get(nc_data_Jan, "latitude", verbose = F)
# Create an array of the SST data
temp.array.Jan <- ncvar_get(nc_data_Jan, "thetao")
temp.array.Feb <- ncvar_get(nc_data_Feb, "thetao")
# Take just the surface layer
surface.temp.Jan <- temp.array.Jan[, , 1] 
surface.temp.Feb <- temp.array.Feb[, , 1] 
# Make it into a raster
SST_Jan <- raster(t(surface.temp.Jan), 
                   xmn=min(nc_lon), 
                   xmx=max(nc_lon), 
                   ymn=min(nc_lat), 
                   ymx=max(nc_lat), 
                   crs=crs("+init=epsg:4326")) %>% 
  flip(., direction = "y")
# Same for Feb
SST_Feb <- raster(t(surface.temp.Feb), 
                   xmn=min(nc_lon), 
                   xmx=max(nc_lon), 
                   ymn=min(nc_lat), 
                   ymx=max(nc_lat), 
                   crs=crs("+init=epsg:4326")) %>% 
  flip(., direction = "y")
# Calculate weighted mean (weighting by the number of observations in January vs February)
wMeanSST <- stack(c(SST_Jan, SST_Feb)) %>% 
  weighted.mean(., w = c(0.64, 0.36))
# Plot for study area
raster::plot(wMeanSST, xlim = c(-27.96598, -24.623), ylim = c(-58.41806, -57.28708), col=viridis(100))
```

The low resolution of this data makes the predictions very pixelated. Interpolate before making predictions.
```{r}
# plot around Saunders
resample(x = wMeanSST, y = dist, method = 'bilinear') %>% plot(xlim = c(-27.96598, -24.623), ylim = c(-58.41806, -57.28708), col=viridis(100))
# interpolate across entire archipelago
wMeanSST_resampled <- resample(x = wMeanSST, y = dist, method = 'bilinear')
```


### Make a 1km x 1km grid  

Use this to sample the distance and SST rasters. The values for the max and min long/lat are from converting the depth raster to LAEA after cropping it to the study area extent (above).
```{r}
long <- seq(-119511.7 , 83876.27, 1000) #first minimum longitude in m, then max, 1000 is 1km
# head(long)
lat <- seq(-49280.75, 112391.2 , 1000)
# head(lat)
Saunders_1KmPoints <- expand.grid(long,lat)
names(Saunders_1KmPoints) <- c("Lon","Lat")
# make it spatial
coordinates(Saunders_1KmPoints) <- ~Lon+Lat
proj4string(Saunders_1KmPoints) <- CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")


# change it back to wgs84 to extract the environmental data (those layers are WGS84)
Saunders_1KmPoints <- spTransform(Saunders_1KmPoints, CRS = proj4string(dist))


# sample the distance and SST layers
Saunders_1KmPoints$SST <- raster::extract(wMeanSST_resampled, Saunders_1KmPoints)
Saunders_1KmPoints$colonydist <- raster::extract(dist, Saunders_1KmPoints)

```

### Make predictions for Saunders

```{r}
Saunders_1KmPoints$GAM_pred <- as.numeric(predict(GAM, type="response", newdata = Saunders_1KmPoints))
```

Make a new raster with the correct resolution and extent of the 1km grid, then use it to convert ```Saunders_1KmPoints``` into a raster for plotting.
```{r}
# new raster in LAEA
r <- raster(ncols = (max(lat) - min(lat)) / 1000, 
            nrows = (max(long) - min(long)) / 1000, 
            crs = CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")) 
# define the extent of the raster
extent(r) <- c(-119511.7 , 83876.27, -49280.75, 80319.25  )
# change points into LAEA as well
Saunders_1KmPoints <- spTransform(Saunders_1KmPoints, CRS = "+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
# rasterize the predicted values
r3 <- rasterize(Saunders_1KmPoints, r, 'GAM_pred', fun=mean)
# project back to WGS84 for plotting
r3 <- projectRaster(from = r3, to = dist)
#plot(r3, col=viridis(100))


# convert to a df for plotting in two steps,
# First, to a SpatialPointsDataFrame
r3_pts <- rasterToPoints(r3, spatial = TRUE)
# Then to a 'conventional' dataframe
r3_df  <- data.frame(r3_pts)
rm(r3_pts)

ggplot() +
  geom_raster(data = r3_df , aes(x = x, y = y, fill = layer)) + 
  scale_fill_viridis() +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "white") +
  ggtitle("Predicted distribution around Saunders Island") +
  coord_fixed(ratio = 1,
                xlim = c(-27.96598, -24.623),
                ylim = c(-58.41806, -57.28708),
                expand = TRUE,
                clip = "on")
```


What is the actual distribution of the data?
```{r}
All <- read.csv("Chick-rearing_trips/All_chick-rearing_trips.csv", stringsAsFactors = FALSE) 
# Make it spatial
coordinates(All) <- ~LON + LAT
proj4string(All) <- CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
# Reproject to WGS84
All <- spTransform(All, CRS = CRS("+proj=longlat +ellps=WGS84"))


ggplot() +
  geom_density_2d_filled(data = as.data.frame(All), aes(x = LON, y = LAT), bins = 20) +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "white") +
  theme_bw() +
  coord_fixed(ratio = 1,
                xlim = c(-27.96598, -24.623),
                ylim = c(-58.41806, -57.28708),
                expand = TRUE,
                clip = "on") +
  ggtitle("Observed distribution around Saunders Island") +
  xlab("Lat") +
  ylab("Long") +
  theme(legend.position = "none")



```

The actual distribution is pretty close to the predicted distribution.

## 2. Predict around all other colonies

Make raster for entire chain of islands using the bathymetry raster as a starting point.

```{r}
SSI_bath_WGS84 <- raster("ssi_geotif/full_ssi18a.tif") %>% 
  projectRaster(., crs=crs("+init=epsg:4326")) %>% 
  crop(., c(-29, -25, -60, -55.75))

# Read in shapefile for land
SSI_WGS84 <- readOGR("Seamask.shp") %>% 
  crop(., c(450000, 1095192, -795043.9, -100000)) %>% 
  spTransform(., crs("+init=epsg:4326"))

# Cut out land
mask <- mask(SSI_bath_WGS84, SSI_WGS84, inverse=F)

# Plot to check
plot(mask, col=viridis(100))

# Set all values to 1
x <- mask
values(x) <- 1
# Cut out land
x <- mask(x, SSI_WGS84, inverse=F)

# Plot to check
plot(x, col=viridis(100))
```

### SST

This is the SST temperature from above, plotted for the whole study area after interpolating it.
```{r}
wMeanSST_resampled <- resample(x = wMeanSST, y = x, method = 'bilinear')
raster::plot(wMeanSST_resampled, col=viridis(100))
```

### Make a grid with a 1x1km resolution to sample the raster layers

First I need to find the extent of the entire island change when it's in LAEA projection.
```{r}
projectRaster(wMeanSST_resampled, crs = CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m"))
```


```{r}
long <- seq(-256872 , 65105.97, 1000) #first minimum longitude in m, then max, 1000 is 1km
# head(long)
lat <- seq(-230376.3, 335023.5, 1000)
# head(lat)
All_1KmPoints <- expand.grid(long,lat)
names(All_1KmPoints) <- c("Lon","Lat")
# make it spatial
coordinates(All_1KmPoints) <- ~Lon+Lat
proj4string(All_1KmPoints) <- CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")


# change it back to wgs84 to extract the environmental data (those layers are WGS84)
All_1KmPoints <- spTransform(All_1KmPoints, CRS = proj4string(dist))
```

I also need to make a new raster with the correct resolution and extent of the study area, which I will use to rasterise the 1km grid with later.
```{r}
# new raster in LAEA
r <- raster(ncols = (max(long) - min(long)) / 1000, 
            nrows = (max(lat) - min(lat)) / 1000, 
            crs = CRS("+proj=laea +lon_0=-26 +lat_0=-58 +units=m")) 
# define the extent of the raster
extent(r) <- c(-256872 , 65105.97, -230376.3, 335023.5)

```

### Read in all colony locations 
```{r}
colonies <- read.csv("Colony_LatLons.csv", header = TRUE)
coordinates(colonies) <- ~Long+Lat
```


### Make functions to predict the distribution and plot

I've added a line to change predicted probabilities to zero if the distance is greater than 100km from the colony. We know from other studies that chick-provision chinstraps don't forage this far from the colony in general, and only one trip was further than 100km from the colony out of the 388 total chick-rearing trips that we recorded in this study.

```{r}

single_island_prediction <- function(colony_name) {
  location <- colonies %>% subset(., str_detect(Colony, colony_name))
  j <- cellFromXY(x, location)
  x[j] <- 2 
  # Create a distance raster from the colony
  dist <- gridDistance(x, origin=2, omit=NA)
  # sample the distance and SST layers
  All_1KmPoints$SST <- raster::extract(wMeanSST_resampled, All_1KmPoints)
  All_1KmPoints$colonydist <- raster::extract(dist, All_1KmPoints)
  # Make predictions
  All_1KmPoints$GAM_pred <- as.numeric(predict(GAM, type="response", newdata = All_1KmPoints))
  # change predictions for points that are more than 100km away to zero, avoiding NAs
  All_1KmPoints[!is.na(All_1KmPoints$GAM_pred) & All_1KmPoints$colonydist > 100000, "GAM_pred"] <- 0  
  # change points into LAEA so that it matches the raster
  All_1KmPoints <- spTransform(All_1KmPoints, CRS = "+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
  # rasterize the predicted values
  r3 <- rasterize(All_1KmPoints, r, 'GAM_pred', fun=mean)
  # project back to WGS84 for plotting
  r3 <- projectRaster(from = r3, to = dist)
}


plot_predicted_distribution <- function(raster, colony_name) {
  # convert to a df for plotting in two steps,
  # First, to a SpatialPointsDataFrame
  r3_pts <- rasterToPoints(raster, spatial = TRUE)
  # Then to a 'conventional' dataframe
  r3_df  <- data.frame(r3_pts)
  rm(r3_pts)
  ggplot() +
    geom_raster(data = r3_df , aes(x = x, y = y, fill = layer)) + 
    scale_fill_viridis() +
    geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "white") +
    ggtitle(paste0("Predicted distribution around ", colony_name)) +
    coord_fixed(ratio = 1)
}

```




### Run for each colony
Plots turned off for now to reduce run time.
```{r}
# colony_codes <- c("CAND", "BRIS")
# predicted_rasters <- purrr::map(colony_codes, ~assign(paste0(colony_codes, "_predicted"), single_island_prediction(colony_codes)))

colony_code <- "ZAV"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Zavodovski")

colony_code <- "VIS"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Visokoi")

colony_code <- "CAND"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Candelmas")

colony_code <- "VIND"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Vindication")

colony_code <- "SAUN"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Saunders")

colony_code <- "MONT"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Montagu")

colony_code <- "BRIS"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Bristol")

colony_code <- "BELL"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Bellingshausen")

colony_code <- "COOK"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Cook")

colony_code <- "THUL"
assign(paste0(colony_code, "_predicted"), single_island_prediction(colony_code))
plot_predicted_distribution(get(paste0(colony_code, "_predicted")), "Thule")
```

### Predict for all colonies at the same time 

I can't just sum up all the individual raster layers and plot, because the distributions around Thule, Cook, and Bellingshausen overlap so much -> have to create a combined distance raster for distances from all colonies and predict from this.

```{r}
# Find the cells where the colonies are in the raster x
j <- cellFromXY(x, colonies)
x[j] <- 2 

# Create a distance raster from all colonies
dist <- gridDistance(x, origin=2, omit=NA)
plot(dist, col = viridis(100))


# sample the distance layers
All_1KmPoints$SST <- raster::extract(wMeanSST_resampled, All_1KmPoints)
All_1KmPoints$colonydist <- raster::extract(dist, All_1KmPoints)

# Make predictions
All_1KmPoints$GAM_pred <- as.numeric(predict(GAM, type="response", newdata = All_1KmPoints))
# change predictions for points that are more than 100km away to zero, avoiding NAs
All_1KmPoints[!is.na(All_1KmPoints$GAM_pred) & All_1KmPoints$colonydist > 100000, "GAM_pred"] <- 0
# change points into LAEA 
All_1KmPoints <- spTransform(All_1KmPoints, CRS = "+proj=laea +lon_0=-26 +lat_0=-58 +units=m")
# rasterize the predicted values
r3 <- rasterize(All_1KmPoints, r, 'GAM_pred', fun=mean)
# project back to WGS84 for plotting
r3 <- projectRaster(from = r3, to = dist)

# Convert to a SpatialPointsDataFrame
r3_pts <- rasterToPoints(r3, spatial = TRUE)
# Then to a dataframe
r3_df  <- data.frame(r3_pts)
rm(r3_pts)

ggplot() +
  geom_tile(data = r3_df , aes(x = x, y = y, fill = layer)) + 
  scale_fill_gradientn(colours=c("#FFFFFFFF","#003366")) +
  # geom_polygon(data = Pelagic_closed.df, aes(x = long, y = lat, group = group), fill = NA, colour = "grey80") +
  geom_polygon(data = NoTake_50km.df, aes(x = long, y = lat, group = group), fill = NA, colour = "grey40") +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "white", colour = "white") +
  ggtitle(paste0("Probability of occurence around all islands")) +
  #coord_fixed(ratio = 1) +
  coord_quickmap() +
  xlab("Longitude") +
  ylab("Latitude")
```
Exporting the raster for Vicky
```{r}
#writeRaster(r3, filename="POccurrenceAllIslands.tif", format="GTiff")
```


With contour lines
```{r}
marmap_dat <- marmap::as.bathy(mask)

autoplot(marmap_dat, geom=c("contour"), coast = FALSE, colour="grey50", size=0.1) + 
  ylab("Latitude") +
  xlab("Longitude") +
  geom_tile(data = r3_df , aes(x = x, y = y, fill = layer)) + 
  scale_fill_gradientn(colours=c("#47897800","#478978FF"), 
                       breaks = c(0.0, 0.25, 0.5, 0.75, 1.0), 
                       labels = c(0.0, 0.25, 0.5, 0.75, 1.0)) +
  # scale_fill_gradientn(colours=c("#00336600","#0033668A")) +
  geom_polygon(data = NoTake_50km.df %>% filter(hole == FALSE), 
               aes(x = long, y = lat, group = group), fill = NA, colour = "#EE7674", linetype = "twodash") +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "grey30", colour = "grey30") +
  #ggtitle(paste0("Probability of occurrence around all islands")) +
  labs(fill = "Probability of occurrence") +
  #coord_fixed() +
  coord_quickmap() +
  xlab("Longitude") +
  ylab("Latitude") +
  theme(panel.background = element_rect(fill = "white", colour = "grey30"))

ggsave(filename = "Figures/Figure4a_PofOccurrence.pdf", dpi = 300, device = "pdf",
       height = 5, width = 5, units = "in")

```




### Calculate weighted distributions for each island

This uses the single island predicted distributions calculated above. 

Calculate the importance of each cell by dividing its value (the probability of occurrence of a penguin) by the sum of the probability of occurences across all cells in that raster, and then multiply by the colony size. This will give the expected number of individuals that would forage in that cell.



```{r}

importance <- function(x) {
  col <- get(paste0(x, "_predicted"))
  col / cellStats(col, stat = "sum")
}

ZAV_expected <- importance("ZAV") * 600000 * 2 # Convey = c. 1 million, Lynch = 600,000, double for number of individuals, not pairs
VIS_expected <- importance("VIS") * 185000 * 2 # Lynch
CAND_expected <- importance("CAND") * 205000 * 2 # Lynch
VIND_expected <- importance("VIND") * 95000 * 2 # Lynch
SAUN_expected <- importance("SAUN") * 155000 *  2# Lynch
MONT_expected <- importance("MONT") * 10000 * 2 # Convey 5000 - 20,000
BRIS_expected <- importance("BRIS") * 15000 * 2 # Convey
BELL_expected <- importance("BELL") * 36000 * 2 # Convey
COOK_expected <- importance("COOK") * 1000 * 2 # Lynch
THUL_expected <- importance("THUL") * 100000 * 2 # Convey

stack <- stack(ZAV_expected, VIS_expected, CAND_expected, VIND_expected, SAUN_expected, MONT_expected, BRIS_expected, BELL_expected, COOK_expected, THUL_expected)
stack_sum <- calc(stack, sum)

#plot(stack_sum, col=viridis(100))

# cellStats(ZAV_predicted, stat = "sum")
# cellStats(ZAV_expected, stat = "sum")
# 
# plot(ZAV_expected)
```

Export for Vicky
```{r}
#writeRaster(stack_sum, filename="WeightedPOccurrenceAllIslands.tif", format="GTiff")
```



Plot weighted distribution.
```{r}
# Convert to a SpatialPointsDataFrame
stack_sum_pts <- rasterToPoints(stack_sum, spatial = TRUE)
# Then to a dataframe
stack_sum_df  <- data.frame(stack_sum_pts)
rm(stack_sum_pts)

# ggplot() +
#   geom_raster(data = stack_sum_df , aes(x = x, y = y, fill = layer)) + 
#   scale_fill_gradientn(colours=c("#FFFFFFFF","#9a004c")) +
#   geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "grey40") +
#   ggtitle(paste0("Weighted probability of occurence around all islands")) +
#   coord_fixed(ratio = 1) +
#   xlab("Longitude") +
#   ylab("Latitude")
```




Plotting with MPA bits and bobs.
```{r}
ggplot() +
  geom_tile(data = stack_sum_df , aes(x = x, y = y, fill = layer)) + 
  scale_fill_gradientn(colours=c("#FFFFFFFF","#8B95C9")) +
  #geom_polygon(data = Pelagic_closed.df, aes(x = long, y = lat, group = group), fill = NA, colour = "grey80") +
  #geom_polygon(data = Benthic_closed.df, aes(x = long, y = lat, group = group), fill = NA, colour = "grey80") +
  geom_polygon(data = NoTake_50km.df, aes(x = long, y = lat, group = group), fill = NA, colour = "grey60") +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "grey40") +
  ggtitle(paste0("Weighted probability of occurence around all islands")) +
  labs(fill = "Penguins per km^2") +
  #coord_fixed() +
  coord_quickmap() +
  xlab("Longitude") +
  ylab("Latitude") +
  xlim(c(-29, -25)) +
  ylim(c(-60, -55.5))
```


```{r}

autoplot(marmap_dat, geom=c("contour"), coast = FALSE, colour="grey50", size=0.1) +
  ylab("Latitude") +
  xlab("Longitude") +
  geom_tile(data = stack_sum_df, aes(x = x, y = y, fill = layer)) +
  scale_fill_gradientn(colours=c("#8B95C900","#8B95C9FF")) +
  # scale_fill_gradientn(colours=c("#00336600","#0033668A")) +
  geom_polygon(data = NoTake_50km.df %>% filter(hole == FALSE), 
               aes(x = long, y = lat, group = group), fill = NA, colour = "#EE7674",
               linetype = "twodash") +
  geom_polygon(data = SSI_polygons.df, aes(x = long, y = lat, group = group), fill = "grey30", colour = "grey30") +
  #ggtitle(paste0("Probability of occurrence around all islands")) +
  labs(fill = "Penguins per km^2") +
  #coord_fixed() +
  coord_quickmap() +
  xlab("Longitude") +
  ylab("Latitude") +
  theme(panel.background = element_rect(fill = "white", colour = "grey30"))

ggsave(filename = "Figures/Figure4a_WeightedPofOccurrence.pdf", dpi = 300, device = "pdf",
       height = 5, width = 5, units = "in")
```
